// src/core/http/generated/models/paymentGateway.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { PaymentGatewayConfigJson } from './paymentGatewayConfigJson.ts';
import type { PaymentGatewayProvider } from './paymentGatewayProvider.ts';
import type { PaymentGatewayStatus } from './paymentGatewayStatus.ts';

export interface PaymentGateway {
  id: string;
  name: string;
  provider: PaymentGatewayProvider;
  status: PaymentGatewayStatus;
  is_sandbox: boolean;
  is_default: boolean;
  public_key?: string;
  access_token?: string;
  client_id?: string;
  client_secret?: string;
  webhook_secret?: string;
  has_public_key?: boolean;
  has_access_token?: boolean;
  has_client_id?: boolean;
  has_client_secret?: boolean;
  has_webhook_secret?: boolean;
  config_json?: PaymentGatewayConfigJson;
  status_message?: string;
  created_at?: string;
  updated_at?: string;
}